django/backend/requirements.txt
asgiref==3.3.4
amqp==5.0.9
celery==5.2.3
channels==3.0.4
daphne==3.0.2
Django==3.2.9
djangorestframework==3.12.4
redis==3.5.3
pytz==2021.3
sqlparse==0.4.1
gunicorn==20.1.0
